+++ server/package.json
@@ -6,18 +6,28 @@
   "scripts": {
     "start": "tsc && node ./dist/index.js",
     "debug": "export DEBUG=* && npm run start",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "nodemon --watch . --ext ts --exec \"mocha -r ts-node/register test/**/*.test.ts\""
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "sqlite3": "^5.0.1"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.14",
     "@types/cors": "^2.8.9",
     "@types/express": "^4.17.11",
+    "@types/mocha": "^8.2.0",
+    "@types/node": "^14.14.22",
+    "@types/sqlite3": "^3.1.6",
+    "chai": "^4.2.0",
+    "chai-http": "^4.3.0",
+    "mocha": "^8.2.1",
+    "nodemon": "^2.0.7",
+    "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
   }
 }
